% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovaTests.R
\name{adonis_repeated_measures}
\alias{adonis_repeated_measures}
\title{Permutational multivariate analysis of variance for repeated measures}
\usage{
adonis_repeated_measures(
  data,
  distmat,
  group1,
  group2,
  sample_id_var,
  rep_meas_var,
  covariates = NA,
  permutations = 999,
  seed = 42,
  first_within = FALSE,
  second_within = FALSE
)
}
\arguments{
\item{data}{Data to use in the test}

\item{distmat}{Distance matrix. Can be a matrix or an object of class
\code{dist}}

\item{group1}{First predictor variable, usually a factor}

\item{group2}{Second predictor variable}

\item{sample_id_var}{Variable that defines the sample IDs, i.e. the
identifiers that correspond to each item in the distance matrix}

\item{rep_meas_var}{Variable that defines the nesting in the experiment,
typically indicating a subject ID or cage ID}

\item{permutations}{Number of permutations}

\item{first_within}{Should the first predictor be shuffled within group?}

\item{second_within}{Should the second fixed effec be shuffled within group?}
}
\value{
The results from \code{vegan::adonis()} in tidy format
}
\description{
Permutational multivariate analysis of variance for repeated measures
}
