% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis-custom.R
\name{adonis_run}
\alias{adonis_run}
\title{Permutational multivariate analysis of variance. Should mostly be used
internally for the permanova_posthoc function}
\usage{
adonis_run(
  data,
  distmat,
  formula_string,
  sample_id_var = SampleID,
  rep_meas_var = NULL,
  permutations = 999,
  seed = 42
)
}
\arguments{
\item{data}{Data to use in the test}

\item{distmat}{Distance matrix, either a matrix or an object of class
\code{dist}}

\item{formula_string}{The right side of the formula to test of class \code{character}}

\item{sample_id_var}{Variable that defines the sample IDs, i.e. the
identifiers that correspond to each item in the distance matrix}

\item{rep_meas_var}{Variable that indicates the repeated measures in the
experiment, typically a subject ID or cage ID. Will be passed to the strata
parameter in the \code{vegan::adonis()} call}

\item{permutations}{Number of permutations}

\item{seed}{Random seed, set just before the initial call to
\code{vegan::adonis()}}
}
\description{
Permutational multivariate analysis of variance. Should mostly be used
internally for the permanova_posthoc function
}
