% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis-custom.R
\name{adonisplus}
\alias{adonisplus}
\title{Permutational multivariate analysis of variance, plus}
\usage{
adonisplus(
  data,
  distmat,
  formula,
  sample_id_var = SampleID,
  rep_meas_var = subject_id,
  shuffle = NULL,
  permutations = 999,
  seed = 42
)
}
\arguments{
\item{data}{Data to use in the test}

\item{distmat}{Distance matrix, either a matrix or an object of class
\code{dist} The distance matrix will automatically be filtered and
re-arranged to match the rows of \code{data}}

\item{formula}{Model formula. The LHS must be "distmat ~". The formula can
either be a literal formula or a string that can be converted into a
formula.}

\item{sample_id_var}{Variable in \code{data} that defines the sample IDs,
i.e. the identifiers that correspond to each item in the distance matrix}

\item{rep_meas_var}{Variable in \code{data} that indicates the repeated
measures in the experiment, typically a subject ID or cage ID}

\item{shuffle}{Named character vector that specifies how to carry out
restricted permutations for variables in \code{data}. Names should
correspond to variables in \code{data}. Values should be either
\code{"between"} or \code{"within"}, meaning that values should be
shuffled between or within levels of \code{rep_meas_var}. See Details for
more info}

\item{permutations}{Number of permutations}

\item{seed}{Random seed, set just before the initial call to
\code{vegan::adonis()}}
}
\value{
The results from \code{vegan::adonis()} in tidy format
}
\description{
Permutational multivariate analysis of variance, plus
}
\details{
A typical experimental design has subjects in a few groups sampled
repeatedly over a few time points. If the variable denoting the group is
\code{study_group} and the variable denoting the time point is
\code{time_point}, then the \code{shuffle} argument would be
\code{c(study_group = "between", time_point = "within")}. During the
permutation stage, the values of the study group will be shuffled between
subjects, preserving the value within each subject. Conversely, the values
of the time point will be shuffled only within each subject.
}
