% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{nmdsplus}
\alias{nmdsplus}
\alias{plot.nmdsplus}
\title{Non-metric multidimensional scaling (NMDS), plus}
\usage{
nmdsplus(data, distmat, sample_id_var = SampleID, num_axes = 2)

\method{plot}{nmdsplus}(x, ...)
}
\arguments{
\item{data}{A data frame giving information on the objects for the NMDS.}

\item{distmat}{A matrix or \code{dist} object giving the distances between
objects in \code{data}. The distance matrix can contain distances between
additional objects not found in \code{data}, but if any of the objects in
\code{data} is missing, an error will occur.}

\item{sample_id_var}{The column in \code{data} that gives the identifiers
for each row, used to find the matching distances in \code{distmat}.}

\item{num_axes}{The number of NMDS axes to return in the output.}

\item{x}{A tibble returned by \code{pcoaplus}.}

\item{...}{Additional aesthetic mappings for ggplot.}
}
\value{
For \code{nmdsplus}, a tibble with new columns giving the position
  of each object along the NMDS axes. We add an additional
  class to the resultant tibble (\code{"nmdsplus"}) to facilitate plotting.
  The attribute "stress" gives the stress in the NMDS.

  For \code{plot.nmdsplus}, a ggplot object.
}
\description{
Non-metric multidimensional scaling (NMDS), plus
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot(nmdsplus)}: Make an NMDS scatter plot

}}
