% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovaTests.R
\name{permanova_with_shuffle_2_groups}
\alias{permanova_with_shuffle_2_groups}
\title{PERMANOVA with 2 group shuffle}
\usage{
permanova_with_shuffle_2_groups(
  dist_matrix,
  s_toTest,
  SampleID,
  group_label1,
  group_label2,
  rep_mes_label,
  covariates,
  perm,
  first_within = F,
  second_within = F
)
}
\arguments{
\item{dist_matrix}{Distance matrix}

\item{s_toTest}{data frame that holds the labels}

\item{SampleID}{column name that holds the IDs}

\item{group_label1}{first fixed effect}

\item{group_label2}{second fixed effect}

\item{rep_mes_label}{random effect}

\item{perm}{number of permutations}

\item{first_within}{should the first fixed effec be shuffled within group?}

\item{second_within}{should the second fixed effec be shuffled within group?}
}
\value{
permuted PERMANOVA results in tidy form
}
\description{
PERMANOVA with 2 group shuffle
}
