% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle_between_groups}
\alias{shuffle_between_groups}
\title{Shuffle a vector, but swap whole groups}
\usage{
shuffle_between_groups(x, g)
}
\arguments{
\item{x}{The vector to be shuffled.}

\item{g}{A vector of groups.}
}
\value{
A version of x where elements have been
  shuffled with each group.
}
\description{
Shuffle a vector, but swap whole groups
}
