% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.adonis}
\alias{tidy.adonis}
\title{Tidy an \code{adonis} object}
\usage{
\method{tidy}{adonis}(x, ...)
}
\arguments{
\item{x}{an object returned from \code{vegan::adonis()}.}

\item{...}{Additional arguments are not used.}
}
\value{
A \code{tibble} with the following columns:
  \item{term}{The name of the regression term.}
  \item{df}{Degrees of freedom used by the model.}
  \item{sumsq}{Sum of squares explained by this term.}
  \item{meansq}{
    Mean sum of squares. Equal to total sum of squares divided by degrees
    of freedom.}
  \item{statistic}{
    The value of a pseudo-F-statistic to use in the permutation test.}
  \item{r.squared}{
    R-squared statistic, or the percent of variation explained by the model.}
  \item{p.value}{P-value from the permutation test.}
}
\description{
Tidy an \code{adonis} object
}
