% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.anova.cca}
\alias{tidy.anova.cca}
\title{Tidy an \code{anova.cca} object produced by \code{adonis2}}
\usage{
\method{tidy}{anova.cca}(x, ...)
}
\arguments{
\item{x}{an object returned from \code{vegan::adonis()}.}

\item{...}{Additional arguments are not used.}
}
\value{
A \code{tibble} with the following columns:
  \item{term}{The name of the regression term.}
  \item{df}{Degrees of freedom used by the model.}
  \item{sumsq}{Sum of squares explained by this term.}
  \item{r.squared}{
    R-squared statistic, or the percent of variation explained by the model.}
  \item{statistic}{
    The value of a pseudo-F-statistic to use in the permutation test.}
  \item{p.value}{P-value from the permutation test.}
}
\description{
Tidy an \code{anova.cca} object produced by \code{adonis2}
}
