% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis-custom.R
\name{adonis_repeated_measures}
\alias{adonis_repeated_measures}
\title{Permutational multivariate analysis of variance for repeated measures}
\usage{
adonis_repeated_measures(
  data,
  distmat,
  group1 = study_group,
  group2 = study_day,
  sample_id_var = SampleID,
  rep_meas_var = subject_id,
  covariates = NA,
  permutations = 999,
  seed = 42,
  group1_permute = "between",
  group2_permute = "within"
)
}
\arguments{
\item{data}{Data to use in the test}

\item{distmat}{Distance matrix, either a matrix or an object of class
\code{dist}}

\item{group1}{First predictor variable, usually a factor}

\item{group2}{Second predictor variable}

\item{sample_id_var}{Variable that defines the sample IDs, i.e. the
identifiers that correspond to each item in the distance matrix}

\item{rep_meas_var}{Variable that indicates the repeated measures in the
experiment, typically a subject ID or cage ID}

\item{covariates}{Any additional covariates to include in the model,
provided as a character vector. If \code{covariates} is not \code{NA}, the
value is inserted directly into the model formula in front of the terms
for \code{group1} and \code{group2}}

\item{permutations}{Number of permutations}

\item{seed}{Random seed, set just before the initial call to
\code{vegan::adonis()}}

\item{group1_permute}{How to perform restricted permutations for
\code{group1}, either \code{"between"} or \code{"within"}. Typically, the
first predictor is a study group, into which each subject is assigned.
In this case, the study group is constant within each subject, and
\code{group1_permute} should be \code{"between"}, to permute the study
groups between subjects.}

\item{group2_permute}{How to perform restricted permutations for
\code{group2}, either \code{"between"} or \code{"within"}. Typically, the
second predictor is a study day, and each subject is sampled over time.
In this case, \code{group2_permute} should be \code{"within"}, to permute
the study days within each subject.}
}
\value{
The results from \code{vegan::adonis()} in tidy format
}
\description{
Permutational multivariate analysis of variance for repeated measures
}
