% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovaTests.R
\name{adonis_repeated_measures}
\alias{adonis_repeated_measures}
\title{Permutational multivariate analysis of variance for repeated measures}
\usage{
adonis_repeated_measures(
  data,
  distmat,
  group1 = study_group,
  group2 = study_day,
  sample_id_var = SampleID,
  rep_meas_var = SubjectID,
  covariates = NA,
  permutations = 999,
  seed = 42,
  group1_within = FALSE,
  group2_within = TRUE
)
}
\arguments{
\item{data}{Data to use in the test}

\item{distmat}{Distance matrix, either a matrix or an object of class
\code{dist}}

\item{group1}{First predictor variable, usually a factor}

\item{group2}{Second predictor variable}

\item{sample_id_var}{Variable that defines the sample IDs, i.e. the
identifiers that correspond to each item in the distance matrix}

\item{rep_meas_var}{Variable that indicates the repeated measures in the
experiment, typically a subject ID or cage ID}

\item{permutations}{Number of permutations}

\item{group1_within}{Does the first predictor change within the unit of
repeated measures? Typically, the first predictor is a study group, into
which each subject is assigned. In this case, the study group is
constant within each subject, and \code{first_within} should be
\code{FALSE}.}

\item{group2_within}{Does the second predictor change within the unit of
repeated measures? Typically, the second predictor is a study day, and
each subject is sampled over time. In this case, the study day assumes
several values within each subject, and \code{group2_within} should be
\code{TRUE}.}
}
\value{
The results from \code{vegan::adonis()} in tidy format
}
\description{
Permutational multivariate analysis of variance for repeated measures
}
