% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{pcoaplus}
\alias{pcoaplus}
\alias{plot.pcoaplus}
\title{Principal coordinates analysis, plus}
\usage{
pcoaplus(data, distmat, sample_id_var = SampleID, num_axes = 2)

\method{plot}{pcoaplus}(x, mapping = ggplot2::aes(x = Axis.1, y = Axis.2), ...)
}
\arguments{
\item{data}{A data frame giving information on the objects for the PCoA.}

\item{distmat}{A matrix or \code{dist} object giving the distances between
objects in \code{data}. The distance matrix can contain distances between
additional objects not found in \code{data}, but if any of the objects in
\code{data} is missing, an error will occur.}

\item{sample_id_var}{The column in \code{data} that gives the identifiers
for each row, used to find the matching distances in \code{distmat}.}

\item{num_axes}{The number of PCoA axes to return in the output.}

\item{x}{A tibble returned by \code{pcoaplus}.}

\item{mapping}{An aesthetic mapping to use for the plot.}

\item{...}{Additional arguments are not used.}
}
\value{
For \code{pcoaplus}, a tibble with new columns giving the position
  of each object along the principal coordinate axes. We add an additional
  class to the resultant tibble (\code{"pcoaplus"}) to facilitate plotting.
  The tibble also contains an attribute, "pctvar", which gives the percent
  variation explained by each principal coordinate axis.

  For \code{plot.pcoaplus}, a ggplot object.
}
\description{
Principal coordinates analysis, plus
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Make a principal coordinates scatter plot
}}

