% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovaTests.R
\name{permanova_with_shuffle_2_groups}
\alias{permanova_with_shuffle_2_groups}
\title{Permutational multivariate analysis of variance with restricted permutations}
\usage{
permanova_with_shuffle_2_groups(
  d,
  data,
  sample_id_col,
  group1,
  group2,
  nesting_var,
  covariates = NA,
  permutations = 999,
  seed = 42,
  first_within = F,
  second_within = F
)
}
\arguments{
\item{d}{Distance matrix. Can be a matrix or an object of class \code{dist}}

\item{data}{Data to use in the test}

\item{sample_id_col}{Column of sample IDs, i.e. the identifiers that
correspond to each item in the distance matrix}

\item{group1}{First predictor variable, usually a factor}

\item{group2}{Second predictor variable}

\item{nesting_var}{Variable that defines the nesting in the experiment,
typically indicating a subject ID or cage ID}

\item{permutations}{Number of permutations}

\item{first_within}{Should the first predictor be shuffled within group?}

\item{second_within}{Should the second fixed effec be shuffled within group?}
}
\value{
The results from \code{vegan::adonis()} in tidy format
}
\description{
Permutational multivariate analysis of variance with restricted permutations
}
